moved query tests
